move sanitize into api common
