+++ components/niagads/api/common/parameters/enums.py
@@ -12,7 +12,7 @@ class EnumParameter(CaseInsensitiveEnum):
     @classmethod
     def validate(cls, value, label: str, returnCls: CaseInsensitiveEnum):
         from niagads.exceptions.core import ValidationError
-        from niagads.utils.string import sanitize  # avoid circular import
+        from niagads.api.common.utils import sanitize  # avoid circular import
 
         try:
             cls(sanitize(value))
