+++ components/niagads/api/common/parameters/location.py
@@ -2,7 +2,7 @@ from fastapi import Query
 from niagads.exceptions.core import ValidationError
 from niagads.genome.core import Assembly, GenomicFeatureType, Human
 from niagads.api.common.models.features.genomic import GenomicFeature
-from niagads.utils.string import sanitize
+from niagads.api.common.utils import sanitize
 
 
 async def assembly_param(
