+++ components/niagads/api/common/parameters/record/path.py
@@ -5,7 +5,7 @@ from fastapi import Path
 from niagads.api.common.parameters.igvbrowser import AnnotatedVariantTrack
 from niagads.genome.core import GenomicFeatureType
 from niagads.api.common.models.features.genomic import GenomicFeature
-from niagads.utils.string import sanitize
+from niagads.api.common.utils import sanitize
 
 
 # tracks and collections
