+++ components/niagads/database/session.py
@@ -3,10 +3,6 @@
 import logging
 import asyncpg
 
-# FIXME: write custom error so that we don't need to import fastapi
-# just for this in everything that uses
-# the database session manager
-from fastapi import HTTPException
 from niagads.exceptions.core import AbstractMethodNotImplemented, ValidationError
 from sqlalchemy.ext.asyncio import (
     create_async_engine,
@@ -88,7 +84,6 @@ class DatabaseSessionManager:
                 AbstractMethodNotImplemented,
                 ValidationError,
                 RuntimeError,
-                HTTPException,
             ) as err:
                 if isinstance(err, AbstractMethodNotImplemented):
                     raise NotImplementedError(str(err))
@@ -104,6 +99,10 @@ class DatabaseSessionManager:
                 raise OSError(f"Database Error: {str(err)}")
 
             except Exception as err:
+                if "HTTPException" in type(err).__name__ or getattr(
+                    err, "status_code", False
+                ):
+                    raise err  # will be handled by API
                 # everything else for which we currently have no handler
                 if "Connection refused" in str(err) or "Connection" in str(type(err)):
                     # don't want to create dependency to redis in this module
