+++ components/niagads/utils/string.py
@@ -11,26 +11,8 @@ import uuid
 from datetime import datetime
 from typing import List, Union
 
-import nh3
 from dateutil.parser import parse as parse_date
 from typing_extensions import deprecated
-
-
-def sanitize(htmlStr: str) -> str:
-    """
-    ammonia sanitization that turns a string into unformatted HTML.
-    used to sanitize incoming API query and path arguments
-
-    Args:
-        htmlStr (str): string to be cleaned
-
-    Returns:
-        str: cleaned string
-    """
-    if htmlStr is not None:
-        return nh3.clean_text(htmlStr.strip())
-
-    return htmlStr
 
 
 def generate_uuid(value: str):
